call-center-server/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
boto3==1.35.39
kafka-python==2.0.2
